Add debug macros and CCI_DEBUG env flag
See the macros at the end of cci_lib_types.h.
To enable them at runtime, export CCI_DEBUG will a comma
separated list (in lowercase) that match the macro definitions.
For example:

  /* enable function entry/exit, memory allocation, and connection
     handling messages. */
  export CCI_DEBUG=func,mem,conn

  /* enables all messages */
  export CCI_DEBUG=all

  /* disables all messages except ERR and WARN */
  unset CCI_DEBUG

Use the debug() macro in the code with the appropriate debug
level. Examples include:

  /* with no arguments */
  debug(CCI_DB_MEM, "allocing rx ring failed");

  /* with an argument */
  debug(CCI_DB_INFO, "duplicate ack %llu", ack);

  /* with two debug levels OR'd together */
  debug(CCI_DB_PEER|CCI_DB_CONN, "closing conn to %s", peer->name);

  /* special functions for entering and exiting functions */
  CCI_ENTER;  /* "entering function foo" */
  CCI_EXIT;   /* "exiting function foo" */

The debug() macro will print on stderr and prepends each message with
"cci: " and adds a newline so that you do not have to put one in the
message.