sock: don't call recvfrom again if it returned eagain
This was part of the lost packet issue. If recvfrom() with PEEK
returned AGAIN, I was calling recvfrom() wthout PEEK to drop
a msg. The race was that a alid msg arrived after the peek and
before the second call. Only call recvfrom to drop if the header
is malformed.