More portals
a) Have to put the portals table index in the ".ini" file.
b) add some minor debug output to cci_accept; can filter later
c) Complete connection algorithm for server side
   -- this is a misnomer for portals; but
      i) we need to honor the connection structure so that
         cci_send will have a target :)
     ii) actually, used a short-circuit to communicate connection
         information between threads; i.e., portals event receiving
         CCI connection request posts a crq; which in turn is
         read by portals_get_conn_request; latter copies it into
         the connection request structure and deletes crq and
         associated memory allocations
    iii) In turn, portals_accept, takes the connection request
         structure and applies data to connection structure;
         cleaning up connection request structure and associated
         memory allocations

At this point, pingpong test server makes connection and cci_send
has appropriate target.  Need to write the portals code to PtlPut
the cci_send message.

Meanwhile, back at the ranch, the client does not yet know that it
connected to the server.  This is due to a disconnect between the
asynchronous portals event processor and the CCI event processor.
Next step is to fix this.