Updated portals_init() to allocate pdev->ep_ids
Initialize pdev->ep_ids{]
Some code cleanup

Currently hangs in portals_get_event()

    -- server-side has no events posted to ep->evts
    -- client-side has same story

This is basically the same problem that I was tracking in the
main branch.  In general, what is happening is that the threaded
implementation requires allocation (and later deallocation) of the
portals events (structure) and the associated md structure.  Otherwise,
the pointer (structure) gets overwritten unintentionally.  In addition,
there is movement which needs to be done within the internal CCI queues
and we need to verify that this is done correctly.

The untracked files do not need to be added.