portals: revamp active message buffering
Create two large AM buffers per endpoint. Set the MD to allow
max_send_size messages. The rx struct now simply carries the
portals event and a pointer to the AM buffer. When the AM buffer
becomes full, we unlink it. When all of the events are returned,
we repost it (PtlMEMDAttach()).

Also, added a token to get_event_ep() to serialize it. This fixed
the portals dropped event messages.