Add endpoint/connection options module
These were not in a group and were skipped. Added a section
for them.

For option names, indicate whether they apply to get only or both
get/set (the orginal comment was ignored).

Get/set_opt() still have "JMS Fill me in" but it is not needed unless
he wants to add more.