portals: support chaning buf count via set_opt
Manage AM buffers differently. Keep lists of active and
orphaned buffers. A call to set_opt to change the rx count
will allocate two new AM buffers. If the old AM buffers are
not being used, free them. Otherwise, add them to the orphan list
until they have no outstanding events.

For free_tx and free_rx, add a force parameter. For endpoint cleanup,
set the force flag. In set_opt(), do not set force and only free
a tx or rx if there is an available idle one. This means that set_opt()
may not reduce the number of buffers to the level that the application
asked for. The app will need to call get_opt() to see what the
final count is.

Factorize creating and posting AM buffers.