Direct connect model
Remove (for now) the connect-to-service model.

Each endpoint has a name that can be passed out-of-band to a
client. The client passes the name to cci_connect() for the
server_uri. There is no port since the connect is directly
to the named endpoint.

In this model, every endpoint can receive connection requests.
The connection request has been converted to a CCI event and
added to the cci_event_t union.

After receiving a connection request event, the application decides
to accept or reject. It then passes the conn_req to either
cci_accept() which returns a new connection or it passes it to
cci_reject(). Both will reliably reply to the client.

The connection request event must still be returned to CCI
using cci_return_event().

Remove service related items