ccieth: start working on get_devices()
If we found devices in the config file, get the one that match the eth driver
Otherwise, take all ethernet interfaces that are up and not loopback in the
output of getifaddrs

That's in get_devices for now, without any global state.
I'll change that when the current get_devices() discussion will be done.