verbs: implement accept()
Pass a header in rdma_connect() with the CCI conn attr and
payload length.
For now, pass payload inline as well, but need to change to
sending after the QP is up to send a full MTU.
Implement part of get_event(). It currently gets CM conn request
events only.