ccieth: report connection in send/recv events
Pass the user connection pointer to the driver in connect and accept
so that it can pass it back to the lib in events.
The user connection is now allocated on connect() instead of when
the ACCEPTED event arrives. We can store the context there instead
of passing it through the driver.