ccieth: when a connect packet points to an invalid endpoint, ack with a new NO ENDPOINT status
If our connect accept is nacked like this (connect initiator closed in the meantime),
our side of the connection is marked as invalid.
If a reject is nacked, we just destroy the rejected connection as usual.
If a request is nacked, we pass a reject to user-space because I don't see anything better in the API...