tests: make the old ccieth debugging a regular test program "loopback"
It builds but does not run during make check because I don't know
if it would work with other drivers.

This loopback test uses a single endpoint, we can't assume whether
the send completion will arrive before or after the recv event,
same for connect done or accept done events.