ccieth: make the recv seqnum bitmap a __u32
We can't portably pass a ulong on the wire (we'll pass the bitmap soon),
and there's no point in marking pkt 32-63 as recvd in we can't ack them.

By the way, optimize the ffz() code since we have a single ulong to consider
and we seem to be sure it will contain some zero.

If __u32 is ever equal to ulong, we'll need to do
  relfull = bitmap == ~0UL ? CCIETH_CONN_RECV_BITMAP_BITS : ffz(bitmap)