multitransport: use the callbacks of the target plugin instead cci_core
Add a plugin pointer in these private struct:
* device (set during init or get_devices)
* endpoint (copied from device during create)
* connection (copied from endpoint during accept/connect)

Add a plugin parameter to plugin methods that don't have
any of these types as a context:
* init
* finalize
* get_devices

Update sock() and template()

rma_deregister() is currently a no-op because its public API has
no device/endpoint/connection. We should either add a endpoint
to the public API and use it to find the plugin method, or
call the method of all plugins.