verbs: remove unneeded assert
If the server sleeps rather than polls on get_event() and then wakes for
a client connection request and it calls accept(), the client
can send a message and the server can get the MSG's completion before
the RDMA_CM_EVENT_ESTABLISHED event. This assert is not needed since
the server's conenction is already fully up.

Reported by Facundo Domínguez, ParSci