API: app passes cci_opt_t to [get|set]_opt
Change cci_opt_handle_t handle to a void *. Drop cci_opt_level_t.
Use the name to determine if the handle is an endpoint or connection.
Create a union of the types that can be passed to these functions.
The app allocates storage for the union.

The downside to this approach is that transports cannot have private
option names.