sock: detect interface availability and information with getifaddrs and ethtool
We get the MTU, speed, UP/DOWN and PCI info on Linux.

On other OSes, we only get UP/DOWN. MTU might work because it uses sockets, not ethtool, to be tried.

This code was stolen from ccieth, that's why the common part goes in src/api/init.c
(near other dev-management functions)

If getifaddrs isn't available, we fallback to the previous behavior
(unknown device info)