verbs: properly handle RO conns
using both RDMA and SendRecv paths.
Use a 16b seqno per RO connection. If a RDMA MSG arrives
early, leave it in place and check it later. If a SendRecv
MSG arrives early queue it on the vconn->early queue. When
reaping a RDMA MSG, check if the early queue is empty. If
not see if the MSGs are next.

Thanks to Dave for suggesting how to simplify the RDMA MSG
header which allows for a larger sequence number. Add 1 to
the len so the header always has a bit set. Subtract the 1
when parsing the header.