verbs: tell valgrind to ignore reading of extra bytes
When using RDMA MSGs, we need to align the length so that the data
will arrive aligned. To do so, we add pad bytes after the MSG and
before the header. We read beyond the MSG buffer which might exceed
the end of the allocated space which valgrind notices. Tell
valgrind it is okay to read these bytes. This only applies for messages
that can be inlined on the sender. It does not apply for larger
MSGs since we copy them into registered buffers that are large
enough for a full MTU MSG and header.