gni: avoid tx cq overrun which is non-recoverable
Double GNI_EP_TX_CNT when creating the tx cq. Use a counter to
track the number of outstanding rma_ops. Increment the counter
when allocing a new rma_op and decrement it when freeing it. If
the count hits the max, return CCI_ENOBUFS (for lack of a better
error).

Ideally, we could create more rma_ops and simply not post them
until we reap rma completion. Or, we could pre-allocate blocks
of rma_ops with a dedicated cq, but we would have to poll many
cqs.