gni: implement fence
To track rma_ops, we have three queues:
  gep->rma_ops holds posted RMAs
  gconn->rma_ops holds queued RMAs waiting on remote handles
  gconn->fenced holds RMAs behind a fence

We use the gep->rma_op_cnt to ensure that we do not overrun the
tx CQ. When it is at the limit, we post one rma_op for each
RMA completion. We use the gconn->rma_op_cnt to let us know when
fenced RMAs can start.

When gconn->rma_op_cnt drops to 0, we check if there are fenced
RMAs. If so, we move them to the gconn->rma_ops list. From there,
they will be posted once they have their remote handle.

We always request remote handles even when fencing to overlap the
request while it sits on the fence.

This does not ensure in-order posting of RMAs on RO connections.
That will need more work.