gni: add blocking support
Uses event fd for the endpoint's fd. We have a flag to indicate if
the fd is empty and we only write to it if it is empty (no previous
events pending) and we only read from it if the app is consuming the
last event on the queue.

We have to create the GNI CDM and CQs in the progress thread to
ensure that the signals will be delivered there. The app thread
never calls and GNI completion functions; only the progress thread
can safely all them.

This code can reliably crash gemini nodes. It may be because it
still launches sends (MSGs and RMAs) from the app thread. I will
investigate queuing sends and having the progress thread do all
sending.