verbs: drain the CQ before rearming CQ
This fixes the case where create_endpoint() creates an OS
handle but the app never uses it (the app polls on
get_event()).

Track when we get an IB event and poll the CQ. Skip checking
for new events until we have completely drained the CQ.

Thanks to Facundo DomC-nguez <facundominguez@gmail.com> for the
test case.