API: change RMA handle from uint64_t
to cci_rma_handle_t

To avoid looking up a remote uint64_t RMA handle to obtain
the necessary data for native RMA interfaces (and the
related issues with caching such lookups), create a
cci_rma_handle_t which contains four uint64_ts for the
transport to use as needed. The application does not need
to inspect the handles and can pass them around. The transport
is resposible for serializing any stored data so that they
can be passed over the network.

This patch updates:

API: header and src/api functions
ctp: transport function signatures
tests: pingpong and register (only RMA tests)
verbs: compiles and runs correctly
sock: compiles but is probably broken

Need to update:
gni
eth
portals
tcp (in topic/tcp)

The verbs transport stores the registered base address in
rma_handle->stuff[0] and the rkey in stuff[1].

The sock transport stuff the address of the sock_rma_handle_t
in rma_handle->stuff[0]. This uint64_t is equivalent to the
old uint64_t handle. I added a couple of FIXME comments in
the sock transport to indicate areas of probable brokeness
(one of which predates this change).