tcp: use correct tx id when sending conn reply
This fixes a bug reported by Rob Stewart. The conn reply contained
the sender's (server's) tx id when it should have been the receiver's
(client's) tx id.

Since the app does not have the connection yet, the socket is empty and
there is no chance of losing a message marker, we can send the reply
from accept() and skip progressing the conn.

Also, add more debug when getting an event.