API: use separate RMA event instead of SEND event
Create a new CCI_EVENT_RMA for completion of cci_rma() operations.
The completion generates this event instead of CCI_EVENT_SEND even
if there is a completion message.

This commit is for discussion and only modifies the header. No changes
have been made to the source code or transports (i.e. do not merge
this).