e2e: fix send completions
If UU or R* with error, retrieve the tx from the context and
return the send event to the app.

Do not queue a reliable send on the pending list if the app
requests blocking.

Do not queue UU events on ep->evts and do not check it in get_event.
Wait for the native send event and handle (mentioned above).

Lastly, set the connection attribute on the passive side when
handling the connection request. :-)