tcp: fix race in RMA handling
When the initiator is multi-threaded, there is a race between queuing
the txs on the conn and then queuing the rma_op on the endpoint if
another thread was progressing the endpoint and received an ack before
the rma_op was enqueued on the endpoint.

Take the endpoint lock before queuing the txs.

Reported by Teng Wang <tzw0019@tigermail.auburn.edu>