gni: decrease connection progress time
We expect most communication to be using uGNI. We only use Sockets
for connection setup. We do not poll connections every time because
we do not expect to find any.

The old value took about 1/2 second. This change lowers that to about
50 ms.

Reported by Teng Wang <tzw0019@tigermail.auburn.edu>