sm: make FIFO headers 8-bytes
to hold connection id, etc.

FIFO will hold MSG/RMA headers only. The sender places the payload in
its own MMAP buffer (one for MSGs and optionally one for RMA if no knem
or CMA). The RMA MMAP buffer will hold parts of the RMA op in the first
cacheline and the payload starts at the next cacheline.

Add the design.txt with the description of what we intend to do.