tcp: RMA completion message wrong offset
When storing the completion message for later retrieval, we copied it
with the wrong offset (or we passed the wrong offset to the send
common function).

Also, if the RMA failed, set the failed status in the tx's send event
status. Before we stored in the rma_op just before freeing it (and
never reporting it to the user).

FIXME: There is still a bug in the progress_rma() if the target cannot
get a tx for the RMA read reply. It returns RNR to the initiator who
should try again, but fails instead.