sm: distinguish MSGs and RMAs in evt->priv
We use evt->priv for sends to store the slot ID. We use it to store the
RMA pointer for RMAs. To distinguish the two uses, set the lowest bit to
indicate a MSG index and shift the index to the left one bit. The RMA
pointer is guaranteed to always be 0 in the lowest bit.