tests: add rpc test
This test mimics the commnuication path of the Mercury RPC library.
Each RPC sends a MSG to the server, the server initiates an RMA with
completion MSG.

By default, all MSGs are max size for the connection. There are flags
to adjust all sizes (request MSG, RMA, and response MSG).