gni: enable CCI on Cray Aries
Cray changed the PTAG and INSTANCE ID going from Gemini to Aries.

Instead of passing in a ptag on Aries, you use the macro
GNI_FIND_ALLOC_PTAG. During init(), check if the NIC is Gemini
or Aries. If Aries, set the gdev->ptag to this macro.

The instance id is 32 bits on Gemini and only 24 bits on Aries.
Ensure all ids fit in 24 bits.