sm: fix double counting in blocking mode
When sending a MSG to a peer who has enabled blocking mode, we write a
byte to their fifo. When processing the message at the receiver, we
write a second byte to the local fifo. Do not write the second byte at
the receiving side.

Reported by Saurabh Gupta <saurabg@gmail.com>