tcp: implement create_endpoint_at()
Convert create_endpoint() to create_endpoint_at(). If service is passed,
convert it to a numeral and set the sin_port. If not and the device port
is set (via the config file), use it. Else, do not set.

Have create_endpoint() call create_endpoint_at() with service set to
NULL.