sock: implement create_endpoint_at()
Convert create_endpoint() to create_endpoint_at(). If service is not
NULL, convert it to a numeral and set sin_port. If NULL and the device
port is set via the config file, use that. Else, do not set.

Have create_endpoint() call create_endpoint_at() with service set to
NULL.