tcp: add cci__evt_t to rma_op
Use its status and context and remove status and context from the
rma_op. When cleaning up the conn, walk tep->rma_ops and remove RMAs for
the conn.

Also add TCP_CTX_RMA to tcp_ctx_t and add that to the rma_op in
preparation for using the cci__evt_t to queue the rma_op on ep->evts.
This will be needed to implement correctly BLOCKING behavior.